server/requirements.txt
asn1crypto==0.24.0
asyncio==3.4.3
beautifulsoup4==4.7.1
boto3==1.9.82
botocore==1.12.82
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
docutils==0.14
firebase-admin==2.15.0
Flask==1.0.2
gcloud==0.17.0
google-api-core==1.7.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-cloud-vision==0.35.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jws==0.1.3
lockfile==0.12.2
MarkupSafe==1.1.0
msgpack==0.6.0
nltk==3.4
numpy==1.16.0
oauth2client==3.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pycryptodome==3.4.3
pymongo==3.7.2
pyOpenSSL==18.0.0
Pyrebase==3.0.27
python-daemon==2.2.0
python-dateutil==2.7.5
python-firebase==1.2
python-jwt==2.0.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
